Column Attachment Justification changed from Minimum to Midpoint. On Center Corrections Flange Width Corrections Attach to Bottom

+++ Revit_Automation/Source/CustomTypes.cs
@@ -30,7 +30,7 @@ namespace Revit_Automation.CustomTypes
         public string strBuildingName;
         public string strDoubleStudType;
         public string strMaterialType;
-        public double dFlangeOfset;
+        public int dFlangeOfset;
         public double dOnCenter;
         public DirectionWithRespectToRoofSlope dirWRTRoofSlope;
         public List<XYZ> gridIntersectionPoints;
